+++ waspc/cli/src/Wasp/Cli/Command/Doctor.hs
@@ -17,7 +17,6 @@ import System.Process (readProcessWithExitCode)
 import qualified Wasp.Generator.ServerGenerator.Common as Server
 import qualified Wasp.Generator.WebAppGenerator.Common as WebApp
 import qualified Wasp.Node.Version as NodeVersion
-import qualified Wasp.Project.Db.Dev.Postgres as Dev.Postgres
 import qualified Wasp.SemanticVersion as SV
 import Wasp.Util (trim)
 import Wasp.Util.GitRev (gitRevDescription)
@@ -51,8 +50,7 @@ checks =
     ("npm", makeToolVersionCheck "npm" NodeVersion.oldestWaspSupportedNpmVersion (ExceptT NodeVersion.getUserNpmVersion)),
     ("Docker", checkDocker >> return "running"),
     makePortCheck "Client" WebApp.defaultClientPort,
-    makePortCheck "Server" Server.defaultServerPort,
-    makePortCheck "Dev database" (fromIntegral Dev.Postgres.defaultDevPort)
+    makePortCheck "Server" Server.defaultServerPort
   ]
   where
     makePortCheck name port =