+++ waspc/e2e-tests/test-outputs/snapshots/kitchen-sink-golden/wasp-app/.wasp/out/sdk/wasp/client/operations/queries/index.ts
@@ -1,88 +1,97 @@
-import { type QueryFor, createQuery } from './core'
-import type { GetTasks_ext } from '../../../server/operations/queries/index.js'
-import type { GetNumTasks_ext } from '../../../server/operations/queries/index.js'
-import type { GetTask_ext } from '../../../server/operations/queries/index.js'
-import type { GetSerializedObjects_ext } from '../../../server/operations/queries/index.js'
-import type { GetTextUppercaseRequests_ext } from '../../../server/operations/queries/index.js'
-import type { GetDate_ext } from '../../../server/operations/queries/index.js'
-import type { GetAnythingNoAuth_ext } from '../../../server/operations/queries/index.js'
-import type { GetAnythingAuth_ext } from '../../../server/operations/queries/index.js'
-import type { GetTrueVoid_ext } from '../../../server/operations/queries/index.js'
-import type { GetAnyNoAuth_ext } from '../../../server/operations/queries/index.js'
-import type { GetAnyAuth_ext } from '../../../server/operations/queries/index.js'
-import type { GetAnyToNumberSpecified_ext } from '../../../server/operations/queries/index.js'
+import { type QueryFor, createQuery } from "./core";
+import type {
+  RegisteredGetTasks,
+  RegisteredGetNumTasks,
+  RegisteredGetTask,
+  RegisteredGetOldestTask,
+  RegisteredGetSerializedObjects,
+  RegisteredGetTextUppercaseRequests,
+  RegisteredGetDate,
+  RegisteredGetAnythingNoAuth,
+  RegisteredGetAnythingAuth,
+  RegisteredGetTrueVoid,
+  RegisteredGetAnyNoAuth,
+  RegisteredGetAnyAuth,
+  RegisteredGetAnyToNumberSpecified,
+} from "../../../server/operations/queries/index";
 
 // PUBLIC API
-export const getTasks: QueryFor<GetTasks_ext> = createQuery<GetTasks_ext>(
-  'operations/get-tasks',
+export const getTasks: QueryFor<RegisteredGetTasks> = createQuery<RegisteredGetTasks>(
+  "operations/get-tasks",
   ['Task'],
 )
 
 // PUBLIC API
-export const getNumTasks: QueryFor<GetNumTasks_ext> = createQuery<GetNumTasks_ext>(
-  'operations/get-num-tasks',
+export const getNumTasks: QueryFor<RegisteredGetNumTasks> = createQuery<RegisteredGetNumTasks>(
+  "operations/get-num-tasks",
   ['Task'],
 )
 
 // PUBLIC API
-export const getTask: QueryFor<GetTask_ext> = createQuery<GetTask_ext>(
-  'operations/get-task',
+export const getTask: QueryFor<RegisteredGetTask> = createQuery<RegisteredGetTask>(
+  "operations/get-task",
   ['Task'],
 )
 
 // PUBLIC API
-export const getSerializedObjects: QueryFor<GetSerializedObjects_ext> = createQuery<GetSerializedObjects_ext>(
-  'operations/get-serialized-objects',
+export const getOldestTask: QueryFor<RegisteredGetOldestTask> = createQuery<RegisteredGetOldestTask>(
+  "operations/get-oldest-task",
+  ['Task'],
+)
+
+// PUBLIC API
+export const getSerializedObjects: QueryFor<RegisteredGetSerializedObjects> = createQuery<RegisteredGetSerializedObjects>(
+  "operations/get-serialized-objects",
   [],
 )
 
 // PUBLIC API
-export const getTextUppercaseRequests: QueryFor<GetTextUppercaseRequests_ext> = createQuery<GetTextUppercaseRequests_ext>(
-  'operations/get-text-uppercase-requests',
+export const getTextUppercaseRequests: QueryFor<RegisteredGetTextUppercaseRequests> = createQuery<RegisteredGetTextUppercaseRequests>(
+  "operations/get-text-uppercase-requests",
   ['UppercaseTextRequest'],
 )
 
 // PUBLIC API
-export const getDate: QueryFor<GetDate_ext> = createQuery<GetDate_ext>(
-  'operations/get-date',
+export const getDate: QueryFor<RegisteredGetDate> = createQuery<RegisteredGetDate>(
+  "operations/get-date",
   [],
 )
 
 // PUBLIC API
-export const getAnythingNoAuth: QueryFor<GetAnythingNoAuth_ext> = createQuery<GetAnythingNoAuth_ext>(
-  'operations/get-anything-no-auth',
+export const getAnythingNoAuth: QueryFor<RegisteredGetAnythingNoAuth> = createQuery<RegisteredGetAnythingNoAuth>(
+  "operations/get-anything-no-auth",
   [],
 )
 
 // PUBLIC API
-export const getAnythingAuth: QueryFor<GetAnythingAuth_ext> = createQuery<GetAnythingAuth_ext>(
-  'operations/get-anything-auth',
+export const getAnythingAuth: QueryFor<RegisteredGetAnythingAuth> = createQuery<RegisteredGetAnythingAuth>(
+  "operations/get-anything-auth",
   [],
 )
 
 // PUBLIC API
-export const getTrueVoid: QueryFor<GetTrueVoid_ext> = createQuery<GetTrueVoid_ext>(
-  'operations/get-true-void',
+export const getTrueVoid: QueryFor<RegisteredGetTrueVoid> = createQuery<RegisteredGetTrueVoid>(
+  "operations/get-true-void",
   [],
 )
 
 // PUBLIC API
-export const getAnyNoAuth: QueryFor<GetAnyNoAuth_ext> = createQuery<GetAnyNoAuth_ext>(
-  'operations/get-any-no-auth',
+export const getAnyNoAuth: QueryFor<RegisteredGetAnyNoAuth> = createQuery<RegisteredGetAnyNoAuth>(
+  "operations/get-any-no-auth",
   [],
 )
 
 // PUBLIC API
-export const getAnyAuth: QueryFor<GetAnyAuth_ext> = createQuery<GetAnyAuth_ext>(
-  'operations/get-any-auth',
+export const getAnyAuth: QueryFor<RegisteredGetAnyAuth> = createQuery<RegisteredGetAnyAuth>(
+  "operations/get-any-auth",
   [],
 )
 
 // PUBLIC API
-export const getAnyToNumberSpecified: QueryFor<GetAnyToNumberSpecified_ext> = createQuery<GetAnyToNumberSpecified_ext>(
-  'operations/get-any-to-number-specified',
+export const getAnyToNumberSpecified: QueryFor<RegisteredGetAnyToNumberSpecified> = createQuery<RegisteredGetAnyToNumberSpecified>(
+  "operations/get-any-to-number-specified",
   [],
 )
 
 // PRIVATE API (used in SDK)
-export { buildAndRegisterQuery } from './core'
+export { buildAndRegisterQuery } from "./core"