+++ waspc/src/Wasp/Generator/SdkGenerator/Common.hs
@@ -6,7 +6,6 @@ import StrongPath
 import qualified StrongPath as SP
 import qualified Wasp.AppSpec.Operation as AS.Operation
 import Wasp.Generator.Common (GeneratedAppDir)
-import Wasp.Generator.ExternalCodeGenerator.Common (GeneratedExternalCodeDir)
 import Wasp.Generator.FileDraft (FileDraft, createTemplateFileDraft)
 import Wasp.Generator.Monad (Generator)
 import Wasp.Generator.Templates (TemplatesDir)
@@ -63,9 +62,6 @@ sdkRootDirInGeneratedAppDir = [reldir|sdk/wasp|]
 sdkTemplatesDirInTemplatesDir :: Path' (Rel TemplatesDir) (Dir SdkTemplatesDir)
 sdkTemplatesDirInTemplatesDir = [reldir|sdk/wasp|]
 
-extSrcDirInSdkRootDir :: Path' (Rel SdkRootDir) (Dir GeneratedExternalCodeDir)
-extSrcDirInSdkRootDir = [reldir|src|]
-
 makeSdkImportPath :: Path Posix (Rel SdkRootDir) File' -> Path Posix (Rel s) File'
 makeSdkImportPath path = (fromJust . parseRelDirP $ sdkPackageName) </> path
 
@@ -75,8 +71,11 @@ clientTemplatesDirInSdkTemplatesDir = [reldir|client|]
 serverTemplatesDirInSdkTemplatesDir :: Path' (Rel SdkTemplatesDir) (Dir ServerTemplatesDir)
 serverTemplatesDirInSdkTemplatesDir = [reldir|server|]
 
-getOperationTypeName :: AS.Operation.Operation -> String
-getOperationTypeName operation = toUpperFirst (AS.Operation.getName operation) ++ "_ext"
+getGenericOperationDefinitionTypeName :: AS.Operation.Operation -> String
+getGenericOperationDefinitionTypeName = toUpperFirst . AS.Operation.getName
+
+getRegisteredOperationTypeName :: AS.Operation.Operation -> String
+getRegisteredOperationTypeName operation = "Registered" ++ getGenericOperationDefinitionTypeName operation
 
 viteDirInSdkTemplatesDir :: Path' (Rel SdkTemplatesDir) (Dir ViteDir)
 viteDirInSdkTemplatesDir = [reldir|client/vite|]